# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tech-Challenge-Oficina API
  version: 0.1.0
  description: API do Sistema Integrado de Atendimento e Execucao de Servicos da oficina.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
    description: Servidor da aplicacao
paths:
...
  /admin/service-orders:
    post:
      operationId: createServiceOrder
      summary: Cria uma Ordem de Servico completa com orcamento calculado
      description: Captura os valores atuais como snapshots e nao altera o estoque nesta etapa.
      tags:
        - Service Orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceOrderRequest'
      responses:
        '201':
          description: Ordem de Servico criada com status received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/service-orders/{serviceOrder}/diagnosis/start:
    parameters:
      - $ref: '#/components/parameters/ServiceOrderId'
    post:
      operationId: startServiceOrderDiagnosis
      summary: Inicia o diagnostico da Ordem de Servico
      description: Altera automaticamente o status de received para in_diagnosis e registra o instante da transicao.
      tags:
        - Service Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Diagnostico iniciado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/service-orders/{serviceOrder}/diagnosis/complete:
    parameters:
      - $ref: '#/components/parameters/ServiceOrderId'
    post:
      operationId: completeServiceOrderDiagnosis
      summary: Conclui o diagnostico e disponibiliza o orcamento
      description: Valida a composicao persistida, altera automaticamente de in_diagnosis para awaiting_approval e registra o instante da disponibilizacao pela API.
      tags:
        - Service Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Diagnostico concluido e orcamento disponivel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/service-orders/{serviceOrder}/cancel:
    parameters:
      - $ref: '#/components/parameters/ServiceOrderId'
    post:
      operationId: cancelServiceOrder
      summary: Cancela uma Ordem de Servico antes da execucao
      tags:
        - Service Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ordem cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  parameters:
...
    ServiceOrderId:
      name: serviceOrder
      in: path
      required: true
      schema:
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Credenciais ou token invalidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationError:
      description: Dados de entrada invalidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Conflito com o estado persistido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    RequestedService:
      type: object
      required:
        - service_id
        - quantity
      properties:
        service_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
    RequestedInventoryItem:
      type: object
      required:
        - inventory_item_id
        - quantity
      properties:
        inventory_item_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
    CreateServiceOrderRequest:
      type: object
      required:
        - customer_document
        - vehicle_id
        - services
        - inventory_items
      properties:
        customer_document:
          type: string
          description: CPF ou CNPJ, com ou sem pontuacao.
          example: 529.982.247-25
        vehicle_id:
          type: integer
        services:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RequestedService'
        inventory_items:
          type: array
          items:
            $ref: '#/components/schemas/RequestedInventoryItem'
...
    ServiceOrderServiceItem:
      type: object
      required:
        - service_id
        - quantity
        - unit_price
        - subtotal
      properties:
        service_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: integer
          format: int64
          minimum: 0
        subtotal:
          type: integer
          format: int64
          minimum: 0
    ServiceOrderInventoryItem:
      type: object
      required:
        - inventory_item_id
        - type
        - quantity
        - unit_price
        - subtotal
      properties:
        inventory_item_id:
          type: integer
        type:
          type: string
          enum:
            - part
            - supply
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: integer
          format: int64
          minimum: 0
        subtotal:
          type: integer
          format: int64
          minimum: 0
    ServiceOrder:
      type: object
      required:
        - id
        - customer_id
        - vehicle_id
        - status
        - received_at
        - diagnosis_started_at
        - awaiting_approval_at
        - services
        - inventory_items
        - total_amount
      properties:
        id:
          type: integer
        customer_id:
          type: integer
        vehicle_id:
          type: integer
        status:
          type: string
          enum:
            - received
            - in_diagnosis
            - awaiting_approval
            - in_execution
            - cancelled
        received_at:
          type: string
          format: date-time
        diagnosis_started_at:
          type:
            - string
            - 'null'
          format: date-time
        awaiting_approval_at:
          type:
            - string
            - 'null'
          format: date-time
        execution_started_at:
          type:
            - string
            - 'null'
          format: date-time
        cancelled_at:
          type:
            - string
            - 'null'
          format: date-time
        tracking_token:
          type:
            - string
            - 'null'
          description: Retornado somente na criacao administrativa.
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderServiceItem'
        inventory_items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderInventoryItem'
        total_amount:
          type: integer
          format: int64
          minimum: 0
          description: Total calculado pelo servidor em centavos.
    ServiceOrderResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/ServiceOrder'
...
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true